Refactor training module into package structure
Split monolithic training.py into organized submodules:
- augmentation.py for data transformations
- preprocessing.py for record loading and splitting
- tasks.py for task configuration
- core.py for main training logic
