+++ src/lisbet/training/__init__.py
@@ -1,0 +1,4 @@
+from .core import train
+
+__all__ = ["train"]
+__doc__ = """Train module for LISBET."""
